chore(engine): Check feature support when executing query
